@@ -5,7 +5,7 @@ description = "Simulation environment for the project AdaNowo"
 authors = ["Ruben Kins <ruben.kins@rwth-aachen.de>", "luis.kutschat@ita.rwth-achen.de",
     "amrita.raychaudhury@ita.rwth-aachen.de"]
 [tool.poetry.dependencies]
-python = "^3.9"
+python = "3.10"
 numpy = "^1.24.2"
 PyYAML = "^6.0"
 hydra-core = "^1.3.2"
