@@ -1,8 +1,6 @@
 from abc import ABC, abstractmethod
+
 import numpy as np
-from typing import List
-from typing import Tuple
-from typing import Dict
 from omegaconf import DictConfig
 
 from src.abstract_base_class.model_interface import AbstractModelInterface
@@ -12,47 +10,42 @@ class AbstractModelWrapper(ABC):
 
     @property
     @abstractmethod
-    def config(self) -> DictConfig:
-        pass
-
-    @property
-    @abstractmethod
     def n_models(self) -> int:
         pass
 
     @property
     @abstractmethod
-    def output_names(self) -> List[str]:
+    def output_names(self) -> list[str]:
         pass
 
     @property
     @abstractmethod
-    def model_names(self) -> Dict[str, str]:
+    def model_names(self) -> dict[str, str]:
         pass
 
     @property
     @abstractmethod
-    def model_props(self) -> Dict[str, any]:
+    def model_props(self) -> dict[str, any]:
         pass
 
     @property
     @abstractmethod
-    def machine_models(self) -> Dict[str, AbstractModelInterface]:
+    def machine_models(self) -> dict[str, AbstractModelInterface]:
         pass
 
     @property
     @abstractmethod
-    def means(self) -> Dict[str, float]:
+    def means(self) -> dict[str, float]:
         pass
 
     @property
     @abstractmethod
-    def vars(self) -> Dict[str, float]:
+    def vars(self) -> dict[str, float]:
         pass
 
     @property
     @abstractmethod
-    def outputs(self) -> Dict[str, float]:
+    def outputs(self) -> dict[str, float]:
         pass
 
     @property
@@ -61,7 +54,7 @@ class AbstractModelWrapper(ABC):
         pass
 
     @abstractmethod
-    def call_models(self, input: Dict[str, float], latent) -> None:
+    def _call_models(self, input_model: dict[str, float], latent) -> None:
         """
         call all machine models to determine mean and variance (predict_y() or predict_f()) from input and assign
         values to properties means and vars.
@@ -69,7 +62,7 @@ class AbstractModelWrapper(ABC):
         pass
 
     @abstractmethod
-    def interpret_model_outputs(self) -> None:
+    def _interpret_model_outputs(self) -> None:
         """
         sample output distributions dependent on corresponding mean and variance and assign values to properties
         outputs and outputs_array.
@@ -77,7 +70,7 @@ class AbstractModelWrapper(ABC):
         pass
 
     @abstractmethod
-    def get_outputs(self, input: Dict[str, float]) -> Tuple[np.array, dict]:
+    def get_outputs(self, input_model: dict[str, float]) -> tuple[np.array, dict]:
         """
         call methods call_models() and interpret_model_outputs() and return properties outputs and outputs_array.
         """
