@@ -1,14 +1,7 @@
 from abc import ABC, abstractmethod
-from omegaconf import DictConfig
-from typing import Dict
 
 
 class AbstractSafetyWrapper(ABC):
-
-    @property
-    @abstractmethod
-    def config(self) -> DictConfig:
-        pass
 
     @property
     @abstractmethod
@@ -16,7 +9,7 @@ class AbstractSafetyWrapper(ABC):
         pass
 
     @abstractmethod
-    def safetyMet(self, controls: Dict[str, float]) -> bool:
+    def safetyMet(self, controls: dict[str, float]) -> bool:
         """
         check if the controls are within safety constraints.
         return this value and assign it to property safetyFlag.
