@@ -1,6 +1,3 @@
-from typing import List
-from typing import Tuple
-from typing import Dict
 import pathlib as pl
 
 import yaml
@@ -45,7 +42,6 @@ class ModelWrapper(AbstractModelWrapper):
                 if output_name in self._model_names.keys():
                     raise KeyError(f'{self._model_names[output_name]} AND {model_name} have {output_name} as output'
                                    'which leads to multiple uses of same key in dictionaries of ModelWrapper')
-
                 else:
                     self._model_names[output_name] = model_name
                     self._model_props[output_name] = properties
@@ -56,7 +52,7 @@ class ModelWrapper(AbstractModelWrapper):
 
                     # load machine model
                     model_class = properties["model_class"]
-                    path_to_pkl = pl.Path(self._config.pathToModels) / (model_name + '.pkl')
+                    path_to_pkl = pl.Path(self._config.pathToModels) / properties["model_path"]
                     if model_class == "SVGP":
                         mdl = model_interface.AdapterSVGP(path_to_pkl, True)
                     elif model_class == "GPy_GPR":
@@ -66,59 +62,57 @@ class ModelWrapper(AbstractModelWrapper):
                     self._machine_models[output_name] = mdl
 
     @property
-    def config(self) -> DictConfig:
-        return self._config
-
-    @property
     def n_models(self) -> int:
         return self._n_models
 
     @property
-    def output_names(self) -> List[str]:
+    def output_names(self) -> list[str]:
         return self._output_names
 
     @property
-    def model_names(self) -> Dict[str, str]:
+    def model_names(self) -> dict[str, str]:
         return self._model_names
 
     @property
-    def model_props(self) -> Dict[str, any]:
+    def model_props(self) -> dict[str, any]:
         return self._model_props
 
     @property
-    def machine_models(self) -> Dict[str, AbstractModelInterface]:
+    def machine_models(self) -> dict[str, AbstractModelInterface]:
         return self._machine_models
 
     @property
-    def means(self) -> Dict[str, float]:
+    def means(self) -> dict[str, float]:
         return self._means
 
     @property
-    def vars(self) -> Dict[str, float]:
+    def vars(self) -> dict[str, float]:
         return self._vars
 
     @property
-    def outputs(self) -> Dict[str, float]:
+    def outputs(self) -> dict[str, float]:
         return self._outputs
 
     @property
     def outputs_array(self) -> np.array:
         return self._outputs_array
 
-    def call_models(self, input: Dict[str, float], latent=False) -> None:
+    def _call_models(self, input_model: dict[str, float], latent=False) -> None:
         if latent:
             for output_name in self._output_names:
-                self._means[output_name], self._vars[output_name] = self._machine_models[output_name].predict_f(input)
+                self._means[output_name], self._vars[output_name] = \
+                    self._machine_models[output_name].predict_f(input_model)
         else:
             for output_name in self._output_names:
-                self._means[output_name], self._vars[output_name] = self._machine_models[output_name].predict_y(input)
+                self._means[output_name], self._vars[output_name] = \
+                    self._machine_models[output_name].predict_y(input_model)
 
-    def interpret_model_outputs(self) -> None:
+    def _interpret_model_outputs(self) -> None:
         for i, output_name in enumerate(self._output_names):
-            self._outputs[output_name] = np.random.normal(self._means[output_name], self._vars[output_name])
+            self._outputs[output_name] = np.random.normal(self._means[output_name], np.sqrt(self._vars[output_name]))
             self._outputs_array[i] = self._outputs[output_name]
 
-    def get_outputs(self, input: Dict[str, float]) -> Tuple[np.array, dict]:
-        self.call_models(input)
-        self.interpret_model_outputs()
+    def get_outputs(self, input_model: dict[str, float]) -> tuple[np.array, dict]:
+        self._call_models(input_model)
+        self._interpret_model_outputs()
         return self._outputs_array, self._outputs
